WIP: Redesign the UI  for a larger sized window.
Originally the UI was designed for a small popup sized window.
This came with its own challenges, but it was assumed that the devs
would prefer interacting with our extension through the popup window.

Now it's clear that the extension popup UX is very annoying as they
close and lose state if something else on the page is closed.

So I'm redesigning the page for a larger window, with more screen real
estate and more freedom.