build: Do not embed the wasm modules into the bundle, load them at runtime
Why: Firefox submission requires that each .js file must be less than 4MB.

We couldn't use the deferred mode (instead of the embedded mode) of the
existing wasm-loader because it uses relative import paths.

When we try to do relative imports in the content script, it resolves
relative to the webpage it's loaded in.

So we wrote a custom wasm loader that resolves absolute URL of the wasm import
using import.meta.url which will be the url of the JS file which is importing it.