Return a plain object instead of a class from wallet.enable()
Fixes: https://github.com/mlabs-haskell/cardano-dev-wallet/issues/17
SundaeSwap and many other dApps seem to be not happy if the value
returned by wallet.enable() is not a plain object.

They could be doing something like Object.keys() which trips up due to
the presence of internal methods present in the dev wallet, or they
could be copying the API object returned by `wallet.enable()` like
`apiCopy = {...api}`.

Not sure of the exact mechanism, but returning a plain object seem to
fix this.